ci: bsim workflow: nice down test images build
The compile script was modified to fully parallelize
all bsim test images builds.
This speed up the build time, but as we are
building ~80 images, the average load on the system
will skyrocket to over 200.
For CI systems we are seeing in quite many cases,
the CI server loosing connection with the CI runner,
while this part of the job is run, and suspect it is
due to this load.

Nice the job down (by the default +10),
as this should alleviate this problem without slowing
down the job.

(cherry picked from commit b83a82882559f0c4a6bc0a64f83be06cdbdf79c1)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: b83a82882559f0c4a6bc0a64f83be06cdbdf79c1
Change-Id: Idee20f56f5fe0bd7f8513fcb2067e91cc1d10fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4789547
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>