scripts: set_assignee: add an option to assign module PRs
Add a -m option to set_assignees to have the script find all unassigned
module PRs for areas that have an assignee and set the assignee on those
PRs.

This should help avoiding forgotten PRs on modules. It's implemented as
a search and meant to run periodically, so we don't have to add a
workflow for this on every individual module.

(cherry picked from commit 9a1f4abd59751998b0f5201048d87479e7153bd5)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: 9a1f4abd59751998b0f5201048d87479e7153bd5
Change-Id: Ia6bb51fbd5fd8dceeae9d57af8230fd7958c82ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4813740
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>