mm: intel_adsp_tlb: Handle address space conversion warnings
There are various call to z_soc_cached_ptr() which returns values
in the cached address space and are stored in temporary variables
that are not marked as in the cached address space. This results
in sparse complaining about discarding the cached address space
attribute. These temporary variables are then passed to other
internal memory management related functions which do not have
the concept of cached address space (as it is currently Xtensa
specific). Because of this, we cannot change the signature of
these functions. Instead, we force a change of address space
when those temporary variables are being assigned to suppress
sparse warnings.

(cherry picked from commit ab0ed577111b117fd0b7d4b47bc28835031c9138)

Original-Reported-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Original-Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
GitOrigin-RevId: ab0ed577111b117fd0b7d4b47bc28835031c9138
Change-Id: I40a29a1005191f4d3400064f740697dff27ef867
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4808550
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>