doc: move puncover down on the page.
The puncover tool does not appear to be maintained anymore and it is
affected by multiple issues. Providing it as the first in the list
is a disservice to users as they might get frustrated and not notice
the `ram_report` tool that actually works.

Move the information at the bottom of the section, and add a warning
about the fact that this is a third party tool that might or might
not be working.

(Also make the link consistent to avoid "here" being a hyperlink, it
adds nothing to the reader either in RST or HTML.)

(cherry picked from commit 52426c284ea04d27ad10b4dbcaca2847b8aa37e6)

Original-Signed-off-by: Diego Elio Pettenò <flameeyes@meta.com>
GitOrigin-RevId: 52426c284ea04d27ad10b4dbcaca2847b8aa37e6
Change-Id: I48a33889b157ed5763b1d47e2754e2fa07836b7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4808572
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>