drivers: serial: esp32_usb: don't use k_usleep in poll_out function
The previous implementation called k_usleep to wait if the fifo was not
empty. This leads to an exception if called from an ISR (e.g. for
for logging).

In addition to that, the k_usleep leads to noticeable interruptions
when printing strings longer than the 64 bytes of the fifo.

With this commit, the function will busy-wait until all characters are
sent or if the timeout is reached. The timeout will only be reached if
no USB host is connected to the port. After the timeout is reached
once, the function will return immediately for subsequent calls
(dropping the characters to be sent) until the USB host is connected
again.

Fixes #60825

(cherry picked from commit 0fbbe02f071b4827d858e3f8b4cc873e1d670a05)

Original-Signed-off-by: Martin Jäger <martin@libre.solar>
GitOrigin-RevId: 0fbbe02f071b4827d858e3f8b4cc873e1d670a05
Change-Id: Ia7b8b9908b707825374a924bacab06969546c651
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4821044
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Yuval Peress <peress@google.com>