tests: interrupt util: Fix for native/posix targets
This utility header was accessing directly the interrupt
controller HW models, but those are not an interface one
can expect to not change, or to be avaliable in different
targets.
Instead we change it to use the APIs we have a contract
about, that is, the ones all POSIX arch boards are
expected to provide.

(cherry picked from commit 9c0b12947e4e3334abf47e92e43a314b6e4aff76)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 9c0b12947e4e3334abf47e92e43a314b6e4aff76
Change-Id: I73b371017eb0b6c08f6aca579fcc56d838ce7629
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4836109
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>