logging: Minor improvement in instance logging
When using instance logging (e.g. LOG_INST_INF) first argument is
a pointer to the logging object associated with an instance. It is
not used if logging is disabled and that can generate 'unused
variable' warning when logging is disabled.

Previously logging object was not created when logging was disabled
That was done to save memory but because of that object could not be
referenced when logging was disabled. Better approach is to create
empty array instead of a logging object. It does not increase memory
usage but can be referenced and potential compilation warning is
suppressed.

(cherry picked from commit 51914443164da8f9a9fab3be1a0e31f5f6c7fab1)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 51914443164da8f9a9fab3be1a0e31f5f6c7fab1
Change-Id: I15b9f1201f18914f0796bce73d2eb8a6f5783504
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4862341
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>