drivers/modem/modem_cellular.c: Improve bus pipe usage
This commit improves the usage of the bus pipe (connected
to UART) to only open it when the modem is actually
powered and ready, not when leaving the idle state. This
ensures the pipe is flushed before sending the init script,
and re-enables the UART driver if it is disabled due to
errors.

While building a test platform based on the nRF9160 and a
Quectel BG95, it was discovered that the nRF9160 correctly
throws UART errors if the RX is enabled while the UART RX
line is low (which was due to the modem being powered
down).

The improvements should also help help remove the
"<wrn> modem_chat: receive buffer overrun" warning which
would occur during startup as the pipe was opened, but
nothing was receiving the data, causing the buffer to
overflow.

(cherry picked from commit 2d2d3ececd0781be6e3ad3cda97d9d44a510dfd0)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarkix123@gmail.com>
GitOrigin-RevId: 2d2d3ececd0781be6e3ad3cda97d9d44a510dfd0
Change-Id: I6859ac421aa812f40fbef2aaa502b8acd1d0451e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4851900
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>