tests: kernel: usage: skip for mr_canhubk3 board
The mr_canhubk3 board enables by default an off-chip watchdog that must
be serviced to avoid triggering a reset and cannot be disabled on a
per-test basis.

test_all_stats_usage assumes the CPU was never idle before the test
starts but this is not the case for mr_canhubk3 because the off-chip
watchdog driver has a thread kicked off during device init that will
conflict with the expected usage stats on this test. So skip this test
for this board.

(cherry picked from commit 738b12428153c6498a8581bce74a7f50ae50d181)

Original-Signed-off-by: Manuel Argüelles <manuel.arguelles@nxp.com>
GitOrigin-RevId: 738b12428153c6498a8581bce74a7f50ae50d181
Change-Id: If6865b2430e1a9749b6bfba6af8b09be6d4cf51b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4872392
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>