modem: chat: Add partial match feature to modem_chat
This commit adds support for partial matches for the modem_chat
module. A match is a combination of an expected response to a
request along with delimiters to use and a handler for the
parsed response.

The usual behavior of the modem_chat script is to continue to
the next step when any expected response is received. The partial
flag indicates that the script should not proceed to the next
step if the response matches the match. This is useful for
commands which respond with an unspecified number of lines,
followed by an "OK". This flag allows the script to essentially
run in a "while" loop until OK is received.

Along with this addition, a more scalable macro for initializing
the modem_chat match struct, MODEM_CHAT_MATCH_INITIALIZER().
Without this macro, we will need 4 different macros to initialize
the 4 variants of a chat_match, and 8 when the next feature is
added...

(cherry picked from commit 840f2e40b570dcd299bc944c03a183950d635a77)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarkix123@gmail.com>
GitOrigin-RevId: 840f2e40b570dcd299bc944c03a183950d635a77
Change-Id: Idecaab920226c4c203f1f588eb985dea6ec66ea2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4860027
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>