tests: code_relocation: exclude mr_canhubk3 board
The mr_canhubk3 board enables by default an off-chip watchdog that must
be serviced to avoid triggering a reset and cannot be disabled on a
per-test basis. The watchdog driver make uses of a semaphore during
device init and on this test the relocation of the kernel sources
produces a fault. So skip this test for this board.

(cherry picked from commit 78cc40f4e525b23ad1b42745c63a5bd5204541ce)

Original-Signed-off-by: Manuel Argüelles <manuel.arguelles@nxp.com>
GitOrigin-RevId: 78cc40f4e525b23ad1b42745c63a5bd5204541ce
Change-Id: Id1a375cb00afd3685ef51d01d0a57b34da46fe19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4872396
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>