drivers: intel: ssp: Revise receive FIFO draining
The receive FIFO needs to be drained in a different way depending when it
is done.
- before start
If the RX FIFO is in overflow state then we must read all the entries out
to empty it (it was after all full).

- before stop
The DMA might be already running to read out data. Check the FIFO level
change in one sample time which gives us the needed information to decide
to wait for another loop for the DMA burst to finish, wait for the DMA to
start it's burst (DMA request was asserted) or drain the FIFO directly.

No need to drain the RX fifo at probe time.

(cherry picked from commit 97bb67d66cd7461b8cbc7fb72c4b6773013d6dd4)

Original-Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
GitOrigin-RevId: 97bb67d66cd7461b8cbc7fb72c4b6773013d6dd4
Change-Id: I4f1aff36b0f159ce059ddf3f058da5be8626e0d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4962210
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>