tests: bsim: Bluetooth: Add BAP unicast disconnect test
The purpose of this test is to verify that the stack has
the correct behavior if the ACL disconnects while one or
more streams are in the streaming state.

The main point it validates is that unicast server device
can restart advertising using a simple k_work in the
disconnected callback.

The number of connections supported by the test suite has
also been reduced to 3, to avoid some seen issues with
scheduling when CONFIG_BT_MAX_CONN=5.

A part of implementing the test for this purpose, the
unicast client and unicast server have been extended
to also advertise and scan for proper ASCS advertising
data.

(cherry picked from commit 01f7d108f3d1982549535f6fe4c7096b188cb3ed)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 01f7d108f3d1982549535f6fe4c7096b188cb3ed
Change-Id: I066da94685cf6e40016c64877f26dd3e5cb65604
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4882923
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>