net: promiscuous: Fix crash in promiscuous mode
When a packet can't be cloned we crash as we try to initialize the
cursor on a nullptr. We should check if we have a valid pointer,
and if we don't we drop the packet along with a warning.

(cherry picked from commit fa0bbaf66cd66843f60a4ea9205e14fc17a6ea78)

Original-Signed-off-by: Andreas Ålgård <aal@ixys.no>
GitOrigin-RevId: fa0bbaf66cd66843f60a4ea9205e14fc17a6ea78
Change-Id: I284049e15ba222783dd724b3baaf8d3f5fb15d76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4875286
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>