Bluetooth: Host: Fix GATT Long Read for EATT
When EATT is established, the value returned from `bt_att_get_mtu` is
not useful to determine the ATT_MTU that applies to a ATT response. This
is because `bt_att_get_mtu` may return the value for a different bearer
than the request is serviced on.

To fix this, the params struct for the GATT read operation is given a
new field that will record the ATT_MTU that applies to this ATT
operation. This value is then used to determine if the GATT long read
operation is concluded.

(cherry picked from commit 9f42c92e2638f0658d169ccb9a50721b74152462)

Original-Fixes: https://github.com/zephyrproject-rtos/zephyr/issues/61741
Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 9f42c92e2638f0658d169ccb9a50721b74152462
Change-Id: I5dccdccb61ead11298d8c393cd00edf7b057a0d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4967340
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>