west: runners: Add support for a common --reset argument
Some of the runners in the tree have been adding their own,
class-specific versions of a switch to instruct the runner to reset or
not the device after flashing.

In order to better support multi-image builds that require more than one
flash operation, introduce a new --reset,--no-reset command-line
parameter that is part of the RunnerCaps so taht this functionality can
be accessed in a standardized manner.

Implementations for the new parameter are provided for the runner
classes that were already configurable in this regard.

(cherry picked from commit 2d38c095a612e38b209484328f4b7408879d116b)

Original-Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>
Original-Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>
GitOrigin-RevId: 2d38c095a612e38b209484328f4b7408879d116b
Change-Id: I5abf391617fc7c058e281ed8676831ba81ba0c0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4905550
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>