shell: backends: telnet: Added support for echo option
Adding support for echo option if telnet commands are supported. This is
useful when the telnet client is in character mode. It allows to use the
arrow keys, ctrl-c and more. Something to keep in mind is that when
character mode is turned on by the client, network traffic is
considerably increased as each typed character is sent over the wire.

enabled.

(cherry picked from commit 724c162821bf70424e76c83630f41dce2105f0eb)

Original-Note: echo mode is only supported if SHELL_TELNET_SUPPORT_COMMAND is
Original-Signed-off-by: David Corbeil <david.corbeil@dynon.com>
GitOrigin-RevId: 724c162821bf70424e76c83630f41dce2105f0eb
Change-Id: Ie05edb3d3b5e7a49c11719f1e321166b98000c8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5004776
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>